建议类 clean code 修改
# message auto-generated for no-merge-commit merge:
merge main into master

建议类 clean code 修改

Created-by: msprofDeveloper

Match-id-2f31c94877a479640e3f490d3444af660521ce62

+++ collector/dvvp/adx/wrapper/hdc_api.cpp
@@ -662,7 +662,7 @@ int IdeCreatePacket(CmdClassT type, IdeString value,
         return IDE_DAEMON_ERROR;
     }
     uint32_t mallocValueLen = valueLen + 1;
-    uint32_t sendLen = sizeof(struct tlv_req) + mallocValueLen;
+    uint32_t sendLen = static_cast<uint32_t>(sizeof(struct tlv_req)) + mallocValueLen;
     IdeStringBuffer sendBuf = static_cast<IdeStringBuffer>(IdeXmalloc(sendLen));
     IDE_CTRL_VALUE_FAILED(sendBuf != nullptr, return IDE_DAEMON_ERROR, "malloc memory failed");
     IdeTlvReq req = (IdeTlvReq)sendBuf;
@@ -676,7 +676,7 @@ int IdeCreatePacket(CmdClassT type, IdeString value,
         return IDE_DAEMON_ERROR;
     }
     *buf = static_cast<IdeMemHandle>(sendBuf);
-    *bufLen = sizeof(struct tlv_req) + valueLen;
+    *bufLen = static_cast<int>(sizeof(struct tlv_req)) + valueLen;
     return IDE_DAEMON_OK;
 }
